Rename default row type
