+++ src/main/java/io/crate/cli/connections/SQLRowType.java
@@ -1,14 +1,16 @@
-package io.crate.cli.gui.common;
+package io.crate.cli.connections;
+
+import io.crate.cli.gui.common.HasKey;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class DefaultRowType extends LinkedHashMap<String, Object> implements HasKey {
+public class SQLRowType extends LinkedHashMap<String, Object> implements HasKey {
 
     private final String key;
 
-    public DefaultRowType(String key, Map<String, Object> attributes) {
+    public SQLRowType(String key, Map<String, Object> attributes) {
         super(attributes);
         this.key = key;
     }
